Convert all remaining JavaScript files to TypeScript
Converts 11 .js/.jsx files to .ts/.tsx across src/ and tests/:
- src/utils/utils.js → .ts (typed function signatures)
- src/components/StatsTab.jsx → .tsx
- src/components/ImportExportTab.jsx → .tsx
- src/components/UI/Header/Header.jsx → .tsx
- src/components/UI/ThemeToggle/ThemeToggle.jsx → .tsx
- src/components/UI/ThemeToggle/ThemeToggle.stories.jsx → .stories.tsx
- src/components/UI/ThemeToggle/index.js → .ts
- src/pages/popup.jsx → .tsx
- tests/mocks/browser.js → .ts (typed MockStorage class)
- tests/utils.test.js → .test.ts
- tests/runTests.js → .ts

Updates all import paths referencing the old .js/.jsx extensions.
All 351 tests pass.
