--- src/components/UI/ThemeToggle/index.js
@@ -1,1 +1,0 @@
-export { ThemeToggle } from './ThemeToggle.jsx';
