+++ src/components/UI/ThemeToggle/index.ts
@@ -1,0 +1,1 @@
+export { ThemeToggle } from './ThemeToggle';
